requirements.txt
streamlit
openai
langchain
pandas
PyPDF2
python-docx
faiss-cpu
tiktoken
